Pulling knn down to prims. Pulling numpy chunking function into prims utils
