+++ cpp/src/knn/knn.hpp
@@ -14,9 +14,14 @@
  * limitations under the License.
  */
 
+#pragma once
+
 #include "common/cumlHandle.hpp"
 
-#include "ml_cuda_utils.h"
+#include "cuda_utils.h"
+
+#include "common/array_ptr.h"
+#include "selection/knn.h"
 
 #include <faiss/gpu/StandardGpuResources.h>
 #include <faiss/gpu/IndexProxy.h>
@@ -28,245 +33,13 @@
 
 #include <iostream>
 
-#pragma once
-
 namespace ML {
 
 	using namespace faiss;
 
-	template<typename T = float, typename Index = int>
-	class ArrayPtr {
-	  public:
-	  T *ptr;
-	  Index N;
-	};
+	class kNN {
 
-
-	/**
-	 * Chunk a single host array up into one or many GPUs (determined by the provided
-	 * list of gpu ids) and fit a knn model.
-	 *
-	 * @param ptr       an array in host memory to chunk over devices
-	 * @param n         number of elements in ptr
-	 * @param devices   array of device ids for chunking the ptr
-	 * @param n_chunks  number of elements in gpus
-	 * @param out       host pointer (size n) to store output
-	 */
-	template<typename T = size_t, typename ARR_T>
-	inline void chunk_to_device(const ARR_T *ptr, int n, int D,
-	    int* devices, ArrayPtr<ARR_T> *output, int n_chunks,
-	    cudaStream_t stream) {
-
-	    size_t chunk_size = MLCommon::ceildiv<size_t>((size_t)n, (size_t)n_chunks);
-
-	    /**
-	     * Initial verification of memory
-	     */
-	    for(int i = 0; i < n_chunks; i++) {
-
-	        int device = devices[i];
-	        T length = chunk_size;
-	        if(length * i >= n)
-	            length = (chunk_size*i)-T(n);
-	        CUDA_CHECK(cudaSetDevice(device));
-	        if(!verify_size(T(length)*T(D), device))
-	            return;
-	    }
-
-	    #pragma omp parallel for
-	    for(int i = 0; i < n_chunks; i++) {
-
-	        int device = devices[i];
-	        CUDA_CHECK(cudaSetDevice(device));
-
-	        T length = chunk_size;
-	        if(length * i >= n)
-	            length = (T(chunk_size)*i)-T(n);
-
-	        ARR_T *ptr_d;
-	        MLCommon::allocate(ptr_d, T(length)*size_t(D));
-	        MLCommon::updateDevice(ptr_d, ptr+(T(chunk_size)*i),
-	                T(length)*T(D), stream);
-
-	        output[i].N = length;
-	        output[i].ptr = ptr_d;
-	    }
-	};
-
-
-	/** Merge results from several shards into a single result set.
-	 * @param all_distances  size nshard * n * k
-	 * @param all_labels     idem
-	 * @param translartions  label translations to apply, size nshard
-	 */
-	template <class C>
-	void merge_tables (long n, long k, long nshard,
-	           float *distances, long *labels,
-	           float *all_distances,
-	           long *all_labels,
-	           long *translations) {
-	  if(k == 0) {
-	    return;
-	  }
-
-	  size_t stride = n * k;
-	  #pragma omp parallel
-	  {
-	    std::vector<int> buf (2 * nshard);
-	    int * pointer = buf.data();
-	    int * shard_ids = pointer + nshard;
-	    std::vector<float> buf2 (nshard);
-	    float * heap_vals = buf2.data();
-	    #pragma omp for
-	    for (long i = 0; i < n; i++) {
-	      // the heap maps values to the shard where they are
-	      // produced.
-	      const float *D_in = all_distances + i * k;
-	      const long *I_in = all_labels + i * k;
-	      int heap_size = 0;
-
-	      for (long s = 0; s < nshard; s++) {
-	        pointer[s] = 0;
-	        if (I_in[stride * s] >= 0)
-	          heap_push<C> (++heap_size, heap_vals, shard_ids,
-	                 D_in[stride * s], s);
-	      }
-
-	      float *D = distances + i * k;
-	      long *I = labels + i * k;
-
-	      for (int j = 0; j < k; j++) {
-	        if (heap_size == 0) {
-	          I[j] = -1;
-	          D[j] = C::neutral();
-	        } else {
-	          // pop best element
-	          int s = shard_ids[0];
-	          int & p = pointer[s];
-	          D[j] = heap_vals[0];
-	          I[j] = I_in[stride * s + p] + translations[s];
-
-	          heap_pop<C> (heap_size--, heap_vals, shard_ids);
-	          p++;
-	          if (p < k && I_in[stride * s + p] >= 0)
-	            heap_push<C> (++heap_size, heap_vals, shard_ids,
-	                   D_in[stride * s + p], s);
-	        }
-	      }
-	    }
-	  }
-	};
-
-	/**
-	 * Search the kNN for the k-nearest neighbors of a set of query vectors
-	 * @param search_items set of vectors to query for neighbors
-	 * @param n        number of items in search_items
-	 * @param res_I      pointer to device memory for returning k nearest indices
-	 * @param res_D      pointer to device memory for returning k nearest distances
-	 * @param k        number of neighbors to query
-	 */
-	void kneighbors(
-	    const cumlHandle &handle,
-	    const ArrayPtr<float> *input, int n_params, int D,
-	    const float *search_items, int n,
-	    long *res_I, float *res_D, int k) {
-
-	  std::vector<long> *id_ranges = new std::vector<long>();
-
-	  int total_n;
-	  int indices;
-
-	  for(int i = 0; i < n_params; i++) {
-	    ArrayPtr<float> params = input[i];
-      indices++;
-      if(i < params.N)
-          id_ranges->push_back(total_n);
-      total_n += params.N;
-	  }
-
-	  float *result_D = new float[k*size_t(n)];
-	  long*result_I = new long[k*size_t(n)];
-
-	  float *all_D = new float[indices*k*size_t(n)];
-	  long *all_I = new long[indices*k*size_t(n)];
-
-	  ASSERT_MEM(search_items, "search items");
-	  ASSERT_MEM(res_I, "output index array");
-	  ASSERT_MEM(res_D, "output distance array");
-
-
-	  #pragma omp parallel
-	  {
-	          #pragma omp for
-	          for(int i = 0; i < indices; i++) {
-
-	            ArrayPtr<float> params = input[i];
-
-	              cudaPointerAttributes att;
-	              cudaError_t err = cudaPointerGetAttributes(&att, params.ptr);
-
-	              if(err == 0 && att.device > -1) {
-
-	                  CUDA_CHECK(cudaSetDevice(att.device));
-	                  CUDA_CHECK(cudaPeekAtLastError());
-
-	                  try {
-
-	                      faiss::gpu::StandardGpuResources gpu_res;
-
-	                      cudaStream_t stream;
-	                      CUDA_CHECK(cudaStreamCreate(&stream));
-
-	                      gpu_res.noTempMemory();
-	                      gpu_res.setCudaMallocWarning(false);
-	                      gpu_res.setDefaultStream(att.device, stream);
-
-	                      bruteForceKnn(&gpu_res,
-	                                  faiss::METRIC_L2,
-	                                  params.ptr,
-	                                  params.N,
-	                                  search_items,
-	                                  n,
-	                                  D,
-	                                  k,
-	                                  all_D+(long(i)*k*long(n)),
-	                                  all_I+(long(i)*k*long(n)));
-
-	                      CUDA_CHECK(cudaPeekAtLastError());
-	                      CUDA_CHECK(cudaStreamSynchronize(stream));
-
-	                      CUDA_CHECK(cudaStreamDestroy(stream));
-
-	                  } catch(const std::exception &e) {
-	                     std::cout << "Exception occurred: " << e.what() << std::endl;
-	                  }
-
-	              } else {
-	                  std::stringstream ss;
-	                  ss << "Input memory for " << &params << " failed. isDevice?=" <<
-	                      att.devicePointer << ", N=" << params.N;
-	                  std::cout << "Exception: " << ss.str() << std::endl;
-	              }
-	          }
-	  }
-
-	  merge_tables<faiss::CMin<float, int>>(long(n), k, indices,
-	      result_D, result_I, all_D, all_I, id_ranges->data());
-
-	  MLCommon::updateDevice(res_D, result_D, k*size_t(n), handle.getImpl().getStream());
-	  MLCommon::updateDevice(res_I, result_I, k*size_t(n), handle.getImpl().getStream());
-
-	  delete all_D;
-	  delete all_I;
-
-	  delete result_D;
-	  delete result_I;
-	};
-
-
-  class kNN {
-
-		std::vector<ArrayPtr<float>> knn_params;
+		std::vector<MLCommon::ArrayPtr> knn_params;
 
 		int total_n;
 		int indices;
@@ -304,7 +77,7 @@ namespace ML {
      * @param input  an array of pointers to data on (possibly different) devices
      * @param N      number of items in input array.
      */
-		void fit(ArrayPtr<float> *input, int N);
+		void fit(MLCommon::ArrayPtr *input, int N);
 
 		/**
 		 * Chunk a host array up into one or many GPUs (determined by the provided
