+++ cpp/src/spectral/spectral.h
@@ -176,7 +176,7 @@ namespace ML {
             MLCommon::allocate(knn_indices, m*n_neighbors);
             MLCommon::allocate(knn_dists, m*n_neighbors);
 
-            ArrayPtr<float> params[1];
+            MLCommon::ArrayPtr params[1];
             params[0].N = m;
             params[0].ptr = X;
 
@@ -331,7 +331,7 @@ namespace ML {
             MLCommon::allocate(knn_indices, m*n_neighbors);
             MLCommon::allocate(knn_dists, m*n_neighbors);
 
-            ArrayPtr<float> params[1];
+            MLCommon::ArrayPtr params[1];
             params[0].N = m;
             params[0].ptr = X;
