+++ cpp/src_prims/common/array_ptr.h
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace MLCommon {
+  struct ArrayPtr {
+    float *ptr;
+    int N;
+  };
+}
