+++ cpp/test/CMakeLists.txt
@@ -33,9 +33,9 @@ if(BUILD_CUML_TESTS)
         ${CUDA_cusolver_LIBRARY}
         ${CUDA_cusparse_LIBRARY}
         ${CUDA_CUDART_LIBRARY}
-        gpufaisslib
         ${CUDA_cusparse_LIBRARY}
         ${CUDA_nvgraph_LIBRARY}
+        gpufaisslib
         faisslib
         ${CUML_CPP_TARGET}
         pthread
@@ -110,6 +110,8 @@ if(BUILD_PRIMS_TESTS)
         ${CUDA_cusolver_LIBRARY}
         ${CUDA_cusparse_LIBRARY}
         pthread
+        gpufaisslib
+        faisslib
         ${ZLIB_LIBRARIES}
     )
