+++ cpp/test/sg/knn_test.cu
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "knn/knn.h"
+#include "knn/knn.hpp"
 #include <vector>
 #include <gtest/gtest.h>
 #include <cuda_utils.h>
@@ -24,7 +24,6 @@
 namespace ML {
 
 using namespace MLCommon;
-
 
 /**
  *
@@ -38,36 +37,36 @@ class KNNTest: public ::testing::Test {
 protected:
 	void basicTest() {
 
-		// Allocate input
-        allocate(d_train_inputs, n * d);
+      // Allocate input
+      allocate(d_train_inputs, n * d);
 
-        // Allocate reference arrays
-        allocate<long>(d_ref_I, n*n);
-        allocate(d_ref_D, n*n);
+      // Allocate reference arrays
+      allocate<long>(d_ref_I, n*n);
+      allocate(d_ref_D, n*n);
 
-        // Allocate predicted arrays
-        allocate<long>(d_pred_I, n*n);
-        allocate(d_pred_D, n*n);
+      // Allocate predicted arrays
+      allocate<long>(d_pred_I, n*n);
+      allocate(d_pred_D, n*n);
 
-        // make testdata on host
-        std::vector<T> h_train_inputs = {1.0, 50.0, 51.0};
-        h_train_inputs.resize(n);
-        updateDevice(d_train_inputs, h_train_inputs.data(), n*d, 0);
+      // make testdata on host
+      std::vector<T> h_train_inputs = {1.0, 50.0, 51.0};
+      h_train_inputs.resize(n);
+      updateDevice(d_train_inputs, h_train_inputs.data(), n*d, 0);
 
-        std::vector<T> h_res_D = { 0.0, 2401.0, 2500.0, 0.0, 1.0, 2401.0, 0.0, 1.0, 2500.0 };
-        h_res_D.resize(n*n);
-        updateDevice(d_ref_D, h_res_D.data(), n*n, 0);
+      std::vector<T> h_res_D = { 0.0, 2401.0, 2500.0, 0.0, 1.0, 2401.0, 0.0, 1.0, 2500.0 };
+      h_res_D.resize(n*n);
+      updateDevice(d_ref_D, h_res_D.data(), n*n, 0);
 
-        std::vector<long> h_res_I = { 0, 1, 2, 1, 2, 0, 2, 1, 0 };
-        h_res_I.resize(n*n);
-        updateDevice<long>(d_ref_I, h_res_I.data(), n*n, 0);
+      std::vector<long> h_res_I = { 0, 1, 2, 1, 2, 0, 2, 1, 0 };
+      h_res_I.resize(n*n);
+      updateDevice<long>(d_ref_I, h_res_I.data(), n*n, 0);
 
-        kNNParams params[1];
-        params[0] = { d_train_inputs, n };
+      ArrayPtr params[1];
+      params[0] = { d_train_inputs, n };
 
-        knn->fit(params, 1);
-        knn->search(d_train_inputs, n, d_pred_I, d_pred_D, n);
-    }
+      knn->fit(params, 1);
+      knn->search(d_train_inputs, n, d_pred_I, d_pred_D, n);
+  }
 
  	void SetUp() override {
 		basicTest();
@@ -88,14 +87,14 @@ protected:
 	int n = 3;
 	int d = 1;
 
-    long *d_pred_I;
-    T* d_pred_D;
+  long *d_pred_I;
+  T* d_pred_D;
 
-    long *d_ref_I;
-    T* d_ref_D;
+  long *d_ref_I;
+  T* d_ref_D;
 
-    cumlHandle handle;
-    kNN *knn = new kNN(handle, d);
+  cumlHandle handle;
+  kNN *knn = new kNN(handle, d);
 };
